Configuration de l'utilisation du token du context d'authorisation de l'utilisateur lors de sa connection

+++ src/components/WorkoutDetails.js
@@ -1,15 +1,24 @@
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext';
+import { useAuthContext } from '../hooks/useAuthContext';
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 
 const WorkoutDetails = ({ workout }) => {
     const { dispatch } = useWorkoutsContext();
+    const { user } = useAuthContext();
 
     const handleClick = async () => {
+        if (!user) {
+            return
+        }
+
         const response = await fetch('/api/workouts/' + workout._id, {
-            method: 'DELETE'
-        })
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
+        });
         const json = await response.json();
 
         if (response.ok) {
@@ -20,9 +29,7 @@ const WorkoutDetails = ({ workout }) => {
         <div className="workout-details">
             <h4>{workout.title}</h4>
             <p><strong>Load (kg): </strong>{workout.load}</p>
-            <p><strong>Number of reps: </strong>{workout.reps}</p>
-            {/* <p>{workout.createdAt}</p>
-            <span onClick={handleClick}>delete</span> */}
+            <p><strong>Reps: </strong>{workout.reps}</p>
             <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
             <span className="material-symbols-outlined" onClick={handleClick}>delete</span>
         </div>
